All the typo fixes!!!! (And translation fixes)
Oh dear, the typos. Many were consistent and the results of this being written by a non-native English speaker "STOPED" vs "STOPPED", but a lot were just mistakes; I'm kind of surprised this library worked, and it almost certainly didn't fully. "Wheel" sometimes became "Wheell" or "Whreel", sometimes consistent, sometimes just in one place.  

Also corrected "formatter" in place of "formater", "raidians" to "radians", "rigth" to "right", and "resistence" to "resistance". 

Finally, in a change affecting using the library in code, changed all references to "bend" to the more correct "turn", and "backward" to "reverse". (The first was a translation error that hurt my brain every time I had to use it, the other just a personal quirk.)